+++ pkg/validation/validators/validators_test.go
@@ -2,6 +2,8 @@ package validators
 
 import (
 	"testing"
+
+	"github.com/editorconfig-checker/editorconfig-checker/pkg/types"
 )
 
 func TestFinalNewlineTrue(t *testing.T) {
@@ -222,27 +224,28 @@ func TestLineEndingCrlf(t *testing.T) {
 }
 
 func TestIndentation(t *testing.T) {
-	if (Indentation("    x", "space", 4)) != nil {
+	params := types.Params{SpacesAfterTabs: false}
+	if (Indentation("    x", "space", 4, params)) != nil {
 		t.Error("Expected correctly indented line to return an nil")
 	}
 
-	if (Indentation("   x", "space", 4)) == nil {
+	if (Indentation("   x", "space", 4, params)) == nil {
 		t.Error("Expected wrong indented line to return an error")
 	}
 
-	if (Indentation("	x", "tab", 0)) != nil {
+	if (Indentation("	x", "tab", 0, params)) != nil {
 		t.Error("Expected correctly indented line to return an nil")
 	}
 
-	if (Indentation("   x", "tab", 0)) == nil {
+	if (Indentation("   x", "tab", 0, params)) == nil {
 		t.Error("Expected wrong indented line to return an error")
 	}
 
-	if (Indentation("	x", "x", 0)) != nil {
+	if (Indentation("	x", "x", 0, params)) != nil {
 		t.Error("Expected unknown indentation to return nil")
 	}
 
-	if (Indentation("   x", "x", 0)) != nil {
+	if (Indentation("   x", "x", 0, params)) != nil {
 		t.Error("Expected unknown indentation to return nil")
 	}
 }
@@ -282,44 +285,71 @@ func TestSpace(t *testing.T) {
 }
 
 func TestTab(t *testing.T) {
-	if Tab("") != nil {
+	spacesAllowed := types.Params{SpacesAfterTabs: true}
+	spacesForbidden := types.Params{SpacesAfterTabs: false}
+
+	if Tab("", spacesForbidden) != nil {
 		t.Error("Expected empty line to return true regardless of parameter")
 	}
 
-	if Tab("x") != nil {
+	if Tab("x", spacesForbidden) != nil {
 		t.Error("Expected line which starts at the beginning to return true")
 	}
 
-	if Tab("	x") != nil {
+	if Tab("	x", spacesForbidden) != nil {
 		t.Error("Expected correctly indented line to return true")
 	}
 
-	if Tab(" x") == nil {
-		t.Error("Expected falsy indented line to return false")
+	if Tab(" x", spacesAllowed) != nil {
+		t.Error("Expected char after space to return nil")
 	}
 
-	if Tab(" *") != nil {
+	if Tab(" *", spacesForbidden) != nil {
 		t.Error("Expected tab indented block comment without indentation to return nil")
 	}
 
-	if Tab(" * some comment") != nil {
+	if Tab(" * some comment", spacesForbidden) != nil {
 		t.Error("Expected tab indented block comment without indentation and with comment to return nil")
 	}
 
-	if Tab("	 *") != nil {
+	if Tab("	 *", spacesForbidden) != nil {
 		t.Error("Expected tab indented block comment with indentation and without comment to return nil")
 	}
 
-	if Tab("	 * some comment") != nil {
+	if Tab("	 * some comment", spacesForbidden) != nil {
 		t.Error("Expected tab indented block comment with indentation and with comment to return nil")
 	}
 
-	if Tab(" */") != nil {
+	if Tab(" */", spacesForbidden) != nil {
 		t.Error("Expected tab indented block comment without indentation in the last line to return nil")
 	}
 
-	if Tab("	 */") != nil {
+	if Tab("	 */", spacesForbidden) != nil {
 		t.Error("Expected tab indented block comment with indentation in the last line to return nil")
+	}
+
+	if Tab("	   bla", spacesAllowed) != nil {
+		t.Error("Expected tab indented and with spaces aligned line to return nil")
+	}
+
+	if Tab("	 bla", spacesAllowed) != nil {
+		t.Error("Expected tab indented and with spaces aligned line to return nil")
+	}
+
+	if (Tab("		  xx", spacesAllowed)) != nil {
+		t.Error("starting with tabs and trailing spaces with at least one character after the spaces")
+	}
+
+	if (Tab("		x", spacesForbidden)) != nil {
+		t.Error("starting with tabs and characters after that")
+	}
+
+	if (Tab("  	a", spacesForbidden)) == nil {
+		t.Error("Starting with two spaces, followed by a tab and followed by a non whitespace char should not return nil")
+	}
+
+	if (Tab(" *", spacesForbidden)) != nil {
+		t.Error("starting with a space and at leat one character after the space")
 	}
 }